test for files historic plugin retention policy

+++ block-node/block-providers/files.historic/src/main/java/org/hiero/block/node/blocks/files/historic/BlocksFilesHistoricPlugin.java
@@ -266,45 +266,50 @@ public final class BlocksFilesHistoricPlugin implements BlockProviderPlugin, Blo
     }
 
     private void cleanup() {
-        final long totalStored = availableBlocks.size();
-        long excess = totalStored - retentionPolicyThreshold;
-        // the numberOfBlocksPerZipFile should generally be immutable once set
-        // for the first time when the block node was originally started
-        // we can rely on the check below to ensure we are deleting the correct
-        // number of blocks
-        while (excess >= numberOfBlocksPerZipFile) {
-            // if we have passed the above check, we can delete at least one zip file
-            // we assume there are no gaps in the zips, the number of blocks per zip file
-            // setting is not possible to change after starting the system for the first time,
-            // also the number of blocks per zip file is a power of ten, so we can safely
-            // say that whatever the range is, it will always start/end with a predictable number
-            // e.g. 0-9, 10-19, 20-29 (batch 10s) or 10_000-19_999, 20_000-29_999 (batch 10_000s) etc.
-            // depending on the setting
-            final long minBlockNumberStored = availableBlocks.min();
-            // no need to compute existing below, we need the path to the zip file, we do not need to
-            // check if the minBlockNumberStored exists, moreover we do not need to know actual block compression type.
-            final Path zipToDelete =
-                    BlockPath.computeBlockPath(config, minBlockNumberStored).zipFilePath();
-            if (Files.exists(zipToDelete)) {
-                try {
-                    // since we keep track of the whole zip file size, that is
-                    // what we should decrement the total bytes stored by
-                    final long zipFileSize = Files.size(zipToDelete);
-                    Files.delete(zipToDelete);
-                    totalBytesStored.addAndGet(-zipFileSize);
-                    // we know that the minBlockNumberStored is for sure the beginning of the batch
-                    // of blocks that we are deleting, also we know that the numberOfBlocksPerZipFile
-                    // is immutable once set originally when we first started the block node,
-                    // so we can safely calculate and remove the range of blocks from the available blocks
-                    availableBlocks.remove(minBlockNumberStored, minBlockNumberStored + numberOfBlocksPerZipFile - 1);
-                } catch (final IOException e) {
-                    throw new UncheckedIOException(e);
-                    // @todo(1268) what to do here if we cannot delete the zip file?
+        // we only take action if the threshold is greater than -1
+        if (retentionPolicyThreshold > -1) {
+            final long totalStored = availableBlocks.size();
+            long excess = totalStored - retentionPolicyThreshold;
+            // the numberOfBlocksPerZipFile should generally be immutable once set
+            // for the first time when the block node was originally started
+            // we can rely on the check below to ensure we are deleting the correct
+            // number of blocks
+            while (excess >= numberOfBlocksPerZipFile) {
+                // if we have passed the above check, we can delete at least one zip file
+                // we assume there are no gaps in the zips, the number of blocks per zip file
+                // setting is not possible to change after starting the system for the first time,
+                // also the number of blocks per zip file is a power of ten, so we can safely
+                // say that whatever the range is, it will always start/end with a predictable number
+                // e.g. 0-9, 10-19, 20-29 (batch 10s) or 10_000-19_999, 20_000-29_999 (batch 10_000s) etc.
+                // depending on the setting
+                final long minBlockNumberStored = availableBlocks.min();
+                // no need to compute existing below, we need the path to the zip file, we do not need to
+                // check if the minBlockNumberStored exists, moreover we do not need to know actual block compression
+                // type.
+                final Path zipToDelete =
+                        BlockPath.computeBlockPath(config, minBlockNumberStored).zipFilePath();
+                if (Files.exists(zipToDelete)) {
+                    try {
+                        // since we keep track of the whole zip file size, that is
+                        // what we should decrement the total bytes stored by
+                        final long zipFileSize = Files.size(zipToDelete);
+                        Files.delete(zipToDelete);
+                        totalBytesStored.addAndGet(-zipFileSize);
+                        // we know that the minBlockNumberStored is for sure the beginning of the batch
+                        // of blocks that we are deleting, also we know that the numberOfBlocksPerZipFile
+                        // is immutable once set originally when we first started the block node,
+                        // so we can safely calculate and remove the range of blocks from the available blocks
+                        availableBlocks.remove(
+                                minBlockNumberStored, minBlockNumberStored + numberOfBlocksPerZipFile - 1);
+                    } catch (final IOException e) {
+                        throw new UncheckedIOException(e);
+                        // @todo(1268) what to do here if we cannot delete the zip file?
+                    }
                 }
+                excess -= numberOfBlocksPerZipFile;
             }
-            excess -= numberOfBlocksPerZipFile;
+            updateGauges();
         }
-        updateGauges();
     }
 
     /**
